create home component
