+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/wales/Desktop/microverse/react/calculator/src/logic/operate.js":"1","/home/wales/Desktop/microverse/react/calculator/src/logic/calculate.js":"2","/home/wales/Desktop/microverse/react/calculator/src/components/App.js":"3","/home/wales/Desktop/microverse/react/calculator/src/components/Button.js":"4","/home/wales/Desktop/microverse/react/calculator/src/index.js":"5","/home/wales/Desktop/microverse/react/calculator/src/helpers/common.js":"6","/home/wales/Desktop/microverse/react/calculator/src/components/Display.js":"7","/home/wales/Desktop/microverse/react/calculator/src/components/ButtonPanel.js":"8"},{"size":487,"mtime":1606667334000,"results":"9","hashOfConfig":"10"},{"size":1922,"mtime":1606667334000,"results":"11","hashOfConfig":"10"},{"size":2005,"mtime":1606669990840,"results":"12","hashOfConfig":"10"},{"size":540,"mtime":1606669985476,"results":"13","hashOfConfig":"10"},{"size":368,"mtime":1606671393695,"results":"14","hashOfConfig":"10"},{"size":176,"mtime":1606667334000,"results":"15","hashOfConfig":"10"},{"size":323,"mtime":1606667555655,"results":"16","hashOfConfig":"10"},{"size":339,"mtime":1606667334000,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"3xqu7c",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/wales/Desktop/microverse/react/calculator/src/logic/operate.js",[],"/home/wales/Desktop/microverse/react/calculator/src/logic/calculate.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/App.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Button.js",[],"/home/wales/Desktop/microverse/react/calculator/src/index.js",[],"/home/wales/Desktop/microverse/react/calculator/src/helpers/common.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Display.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/ButtonPanel.js",[]]
+[{"/home/wales/Desktop/microverse/react/calculator/src/logic/operate.js":"1","/home/wales/Desktop/microverse/react/calculator/src/logic/calculate.js":"2","/home/wales/Desktop/microverse/react/calculator/src/components/Button.js":"3","/home/wales/Desktop/microverse/react/calculator/src/index.js":"4","/home/wales/Desktop/microverse/react/calculator/src/helpers/common.js":"5","/home/wales/Desktop/microverse/react/calculator/src/components/Display.js":"6","/home/wales/Desktop/microverse/react/calculator/src/components/ButtonPanel.js":"7","/home/wales/Desktop/microverse/react/calculator/src/components/Calculator.js":"8","/home/wales/Desktop/microverse/react/calculator/src/components/App.js":"9","/home/wales/Desktop/microverse/react/calculator/src/components/Home.js":"10"},{"size":487,"mtime":1606667334000,"results":"11","hashOfConfig":"12"},{"size":1922,"mtime":1606667334000,"results":"13","hashOfConfig":"12"},{"size":540,"mtime":1606669985476,"results":"14","hashOfConfig":"12"},{"size":231,"mtime":1606672237806,"results":"15","hashOfConfig":"12"},{"size":176,"mtime":1606667334000,"results":"16","hashOfConfig":"12"},{"size":323,"mtime":1606667555655,"results":"17","hashOfConfig":"12"},{"size":339,"mtime":1606667334000,"results":"18","hashOfConfig":"12"},{"size":1963,"mtime":1606671671624,"results":"19","hashOfConfig":"12"},{"size":716,"mtime":1606672888345,"results":"20","hashOfConfig":"12"},{"size":1645,"mtime":1606673343258,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"3xqu7c",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/wales/Desktop/microverse/react/calculator/src/logic/operate.js",[],"/home/wales/Desktop/microverse/react/calculator/src/logic/calculate.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Button.js",[],"/home/wales/Desktop/microverse/react/calculator/src/index.js",[],"/home/wales/Desktop/microverse/react/calculator/src/helpers/common.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Display.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/ButtonPanel.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Calculator.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/App.js",[],"/home/wales/Desktop/microverse/react/calculator/src/components/Home.js",[]]
