+++ src/components/App.js
@@ -1,86 +1,25 @@
-/* eslint-disable no-unused-vars */
-import React, { useState, Component, Fragment } from 'react';
-import ButtonPannel from './ButtonPanel';
-import Display from './Display';
-import Button from './Button';
-import calculate from '../logic/calculate';
+import React from 'react';
+import {
+  Route, BrowserRouter as Router, Switch, Link,
+} from 'react-router-dom';
+import Calculator from './Calculator';
+import Home from './Home';
 
-const buttons = [
-  ['AC', '+/-', '%', '/'],
-  ['9', '8', '7', '*'],
-  ['6', '5', '4', '+'],
-  ['3', '2', '1', '-'],
-  ['0', '.', '='],
-];
-class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      total: null,
-      prev: '0',
-      next: '0',
-      operator: null,
-    };
-  }
-
-  handledisplay = (prev, next, tot, op) => {
-    let ans;
-    if (tot) {
-      ans = tot;
-    } else if (!op) {
-      ans = prev;
-    } else {
-      ans = next;
-    }
-    return ans;
-  };
-
-  handleClick = e => {
-    const {
-      total, next, operator, prev,
-    } = this.state;
-    const res = calculate({
-      total, next, operator, prev,
-    }, e.target.value);
-
-    this.setState({
-      ...res,
-    });
-  };
-
-  render() {
-    const {
-      prev, next, total, operator,
-    } = this.state;
-    return (
-      <>
-        <div className="App">
-          <div className="App__calculator">
-            <Display current={this.handledisplay(prev, next, total, operator)} />
-            <ButtonPannel>
-              {buttons.map(group => (
-                <div key={group} className="row">
-                  {group.map(button => {
-                    const operators = ['+', '-', '*', '/', '='];
-                    const isOperator = operators.includes(button);
-                    return (
-                      <Button
-                        onClick={this.handleClick}
-                        key={button}
-                        name={button}
-                        color={isOperator}
-                        wide={button === '0'}
-                      />
-                    );
-                  })}
-                </div>
-              ))}
-            </ButtonPannel>
-          </div>
+export default function App() {
+  return (
+    <Router>
+      <div className="nav">
+        <h1>Math Magicians</h1>
+        <div className="nav-links">
+          <Link to="/" className="nav-link">Home</Link>
+          <Link to="/calculator" className="nav-link">Calculator</Link>
+          <Link to="/quote" className="nav-link">Quote</Link>
         </div>
-      </>
-    );
-  }
+      </div>
+      <Switch>
+        <Route exact path="/" component={Home} />
+        <Route exact path="/calculator" component={Calculator} />
+      </Switch>
+    </Router>
+  );
 }
-
-export default App;
