+++ src/components/Home.js
@@ -1,0 +1,12 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+export default function Home() {
+  return (
+    <div className="home">
+      <h1>Welcome to our page!</h1>
+      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Officia facilis quo, deleniti at debitis commodi neque eaque dolore! Dolorem voluptatum obcaecati ipsum ratione voluptatibus consequuntur dolor sequi magnam molestiae voluptas nisi iure hic, cumque officia asperiores aspernatur dolores fugiat pariatur consectetur reiciendis quis voluptatem dicta nemo rem. Dolor nobis quia magnam libero ea, impedit mollitia voluptas dolores maxime? Recusandae ratione quas molestiae cupiditate placeat deserunt temporibus velit suscipit pariatur laborum. Veritatis, et ratione accusantium delectus repellat cupiditate voluptas tempore laboriosam amet facere rem omnis perferendis veniam magni dignissimos hic consectetur dolorum aliquid non est qui provident ex? Iusto, eos delectus!</p>
+      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Commodi quas autem harum voluptatibus officia doloremque? Consectetur aut suscipit quas excepturi nisi consequuntur distinctio explicabo sunt, voluptatum rerum eaque quisquam provident inventore assumenda impedit. Quam saepe delectus voluptate tenetur veniam repudiandae similique ipsum eius aspernatur, ut dolorum cupiditate maxime, ratione, voluptatibus ab nam dolorem temporibus expedita deserunt quae. Numquam, expedita autem commodi fugit necessitatibus repudiandae. Deserunt iure accusantium dolorum aliquam voluptatum. Ipsam aperiam aspernatur numquam corporis culpa doloribus commodi ex impedit.</p>
+    </div>
+  );
+}
