+++ src/index.css
@@ -24,7 +24,7 @@ body {
 
 .App {
   width: 100%;
-  height: 100vh;
+  height: 91vh;
   background-color: #fff;
   display: flex;
   align-items: center;
@@ -86,4 +86,32 @@ body {
 
 .wide {
   flex: 1;
+}
+
+.home {
+  padding: 20px;
+  color: black;
+  font-family: sans-serif;
+}
+
+.home p {
+  margin: 20px 0;
+}
+
+.nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
+}
+
+.nav-link {
+  padding: 0 10px;
+  border-right: 2px solid #000;
+  font-weight: bold;
+  font-size: 1.4em;
+}
+
+.nav-link:last-child {
+  border-right: none;
 }
